Create draft performance incidents
We can know in advance of the Grand Final what incidents a performance
is likely to trigger. From the metadata, we can predict a language
change and a singer from another country; and from watching semi-final
performances, we can spot the rest.

The control panel now separates performance incidents into three
groups: those that have happened (you can't duplicate within a
single performance), those that are likely to happen, and those that
haven't happened (but still could).